Fixing some funky terminal issues

+++ linux_env/profile.d/.bash_profile
@@ -4,6 +4,15 @@ PATH=$PATH:$HOME/bin:$HOME/bin/tools:/usr/bin:/usr/sbin:/usr/local/bin:/bin:/sbi
 export PERL5LIB=/data/ops/lib/:$HOME/space/perl
 export PYTHONPATH=$HOME/space/python
 alias jsoncheck='python -mjson.tool'
+declare -A fgcolor
+declare -A bgcolor
+fgchoice=""
+bgchoice=""
+fgcolor=( [black]=30 [red]=31 [green]=32 [yellow]=33 [blue]=34 [magenta]=35 [cyan]=36 [white]=37 )
+bgcolor=( [black]=40 [red]=41 [green]=42 [yellow]=43 [blue]=44 [magenta]=45 [cyan]=46 [grey]=47 )
+defbg=grey
+deffg=black
+
 
 # hide a file
 function dot() {
@@ -41,16 +50,43 @@ function undot() {
 
 function bootstrap() {
   if [[ ! -d "$HOME/git/space" ]]; then
-    mkdir -p "$HOME/git/space"
-    cd $HOME/git
-    git clone http://github.com/rprashad/space space
+    mkdir -p "$HOME/git/space" 2> /dev/null
+    if [[ $? == 0 ]] ; then
+      cd $HOME/git
+      git clone http://github.com/rprashad/space space
+    fi
+  else
+    echo "bootstrap error ocurred"
   fi
 }
 
+
 function setps1() {
+  fgchoice=$1
+  bgchoice=$2
+
+  if [[ -z $fgchoice ]]; then
+    fgchoice=$deffg
+  else if [[ -z ${fgcolor[$fgchoice]} ]]; then
+    echo "fg choice '$fgchoice' - does not exist, setting default: $deffg"
+    fgchoice=$deffg
+  fi
+  fi
+
+  if [[ -z $bgchoice ]]; then
+    bgchoice=$defbg
+  else if [[ -z ${bgcolor[$bgchoice]} ]]; then
+    echo "bg choice '$bgchoice' - does not exist, setting default: $defbg"
+    bgchoice=$defbg
+  fi
+  fi
+
+   echo "BG: $bgchoice: FG: $fgchoice"
+
   case "$TERM" in
   linux|screen|xterm*|rxvt*)
-      PS1="\[\033[34m\][\$(date +%H%M)][\u@\h:\w]$ "
+       # PS1="\[\033[${fgcolor[${fgchoice}]}m\][\$(date +%H%M)][\u@\h:\w]$ "
+	PS1="\[\033[${fgcolor[${fgchoice}]}m\033[${bgcolor[${bgchoice}]}m\][\$(date +%H%M)][\u@\h:\w]$ "
       ;;
   *)
       ;;
@@ -89,14 +125,18 @@ function syncfile() {
   SYNC_DEST=$2
   BAK="$SYNC_DEST.bak"
 
-  # echo "SRC: $SYNC_SOURCE DST: $SYNC_DEST BAK: $BAK"
-  
-  md5_src=`md5sum $SYNC_SOURCE | awk '{print \$1}'`;
-  md5_dst=`md5sum $SYNC_DEST 2> /dev/null | awk '{print \$1}'`;
-  if [[ "$md5_src" != "$md5_dst" ]]; then
-    mv $SYNC_DEST $BAK 2> /dev/null
-    echo "Syncing $SYNC_SOURCE -> $SYNC_DEST"
-    ln $SYNC_SOURCE $SYNC_DEST
+  if [[ -e $SYNC_SOURCE ]]; then
+    # echo "SRC: $SYNC_SOURCE DST: $SYNC_DEST BAK: $BAK"
+    md5_src=`md5sum $SYNC_SOURCE 2> /dev/null | awk '{print \$1}'`;
+    md5_dst=`md5sum $SYNC_DEST 2> /dev/null | awk '{print \$1}'`;
+
+    if [[ "$md5_src" != "$md5_dst" ]]; then
+      mv $SYNC_DEST $BAK 2> /dev/null
+      echo "Syncing $SYNC_SOURCE -> $SYNC_DEST"
+      ln $SYNC_SOURCE $SYNC_DEST
+    fi
+  else
+    echo "Not syncing source $SYNC_SOURCE does not exist"
   fi
 } # syncfile
 
@@ -114,6 +154,27 @@ function synctools() {
      fi
   done
 } # synctools
+
+function allcolors() {
+  # stolen shamelessly from 
+  # http://www.tldp.org/HOWTO/Bash-Prompt-HOWTO/x329.html
+
+  T='gYw'   # The test text
+  echo -e "\n                 40m     41m     42m     43m\
+       44m     45m     46m     47m";
+
+  for FGs in '    m' '   1m' '  30m' '1;30m' '  31m' '1;31m' '  32m' \
+             '1;32m' '  33m' '1;33m' '  34m' '1;34m' '  35m' '1;35m' \
+             '  36m' '1;36m' '  37m' '1;37m';
+    do FG=${FGs// /}
+    echo -en " $FGs \033[$FG  $T  "
+    for BG in 40m 41m 42m 43m 44m 45m 46m 47m;
+      do echo -en "$EINS \033[$FG\033[$BG  $T  \033[0m";
+    done
+    echo;
+  done
+  echo
+} # allcolors
 
 function syncall() {
 # sync tools dir
